Convert EuiToken to TypeScript
