+++ src/components/token/index.ts
